+++ functions/public/Get-FunctionName.ps1
@@ -1,6 +1,7 @@
 Function Get-FunctionName {
     [cmdletbinding()]
-    [outputType("string","PSFunctionName")]
+    [alias('gfn')]
+    [outputType("string", "PSFunctionName")]
     Param(
         [Parameter(
             Position = 0,
@@ -9,21 +10,22 @@ Function Get-FunctionName {
             HelpMessage = "Specify the .ps1 or .psm1 file with defined functions."
         )]
         [ValidateScript({
-            If (Test-Path $_ ) {
-                $True
-                If ($_ -match "\.ps(m)?1$") {
+                If (Test-Path $_ ) {
                     $True
+                    If ($_ -match "\.ps(m)?1$") {
+                        $True
+                    }
+                    Else {
+                        Throw "The path must be to a .ps1 or .psm1 file."
+                        $False
+                    }
                 }
                 Else {
-                    Throw "The path must be to a .ps1 or .psm1 file."
+                    Throw "Can't validate that $_ exists. Please verify and try again."
                     $False
                 }
-            }
-            Else {
-                Throw "Can't validate that $_ exists. Please verify and try again."
-                $False
-            }
-        })]
+            })]
+        [alias("pspath")]
         [string]$Path,
         [Parameter(HelpMessage = "List all detected function names.")]
         [switch]$All,
@@ -31,33 +33,41 @@ Function Get-FunctionName {
         [switch]$Detailed
     )
 
-    $Path = Convert-Path -Path $path
-    Write-Verbose "Parsing $path for functions."
-    $AST = _getAst $path
+    Begin {
+        Write-Verbose "[$((Get-Date).TimeofDay) BEGIN  ] Starting $($myinvocation.mycommand)"
+    } #begin
+    Process {
+        $Path = Convert-Path -Path $path
+        Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Parsing $path for functions."
+        $AST = _getAst $path
 
-    #parse out functions using the AST
-    $functions = $ast.FindAll({ $args[0] -is [System.Management.Automation.Language.FunctionDefinitionAst] }, $true)
-    if ($functions.count -gt 0) {
-        if ($All) {
-            $out = $Functions.Name
-        }
-        Else {
-            $out = $functions.Name | Test-FunctionName
-        }
-        if ($Detailed) {
-            foreach ($item in $($out |Sort-object)) {
-                [pscustomobject]@{
-                    PSTypeName = "PSFunctionName"
-                    Name = $item
-                    Path = $Path
+        #parse out functions using the AST
+        $functions = $ast.FindAll({ $args[0] -is [System.Management.Automation.Language.FunctionDefinitionAst] }, $true)
+        if ($functions.count -gt 0) {
+            if ($All) {
+                $out = $Functions.Name
+            }
+            Else {
+                $out = $functions.Name | Test-FunctionName
+            }
+            if ($Detailed) {
+                foreach ($item in $($out | Sort-Object)) {
+                    [pscustomobject]@{
+                        PSTypeName = "PSFunctionName"
+                        Name       = $item
+                        Path       = $Path
+                    }
                 }
+            }
+            else {
+                $out
             }
         }
         else {
-            $out
+            Write-Warning "No PowerShell functions detected in $path."
         }
-    }
-    else {
-        Write-Warning "No PowerShell functions detected in $path."
-    }
-}
+    } #process
+    End {
+        Write-Verbose "[$((Get-Date).TimeofDay) END    ] Ending $($myinvocation.mycommand)"
+    } #end
+} #close function