+++ functions/public/Test-FunctionName.ps1
@@ -1,5 +1,6 @@
 Function Test-FunctionName {
     [CmdletBinding()]
+    [alias('tfn')]
     [OutputType("boolean", "string")]
     Param(
         [Parameter(Position = 0, Mandatory, ValueFromPipeline, HelpMessage = "Specify a function name.")]