+++ samples/BuildModule.ps1
@@ -1,3 +1,4 @@
+#requires -version 7.1
 
 #build a new module from files
 $files = "$PSScriptRoot\tools.psm1","$PSScriptRoot\Get-ZeroSize.ps1"